Avoid writing a .dart.deferredCount in source_gen mode
We allow writing arbitrary assets but this can't be used because it isn't
declared as an output. We don't run the DeferredRewriter outside of barback so
in source_gen mode it isn't useful to write it at all.

Keeping the outputs to only those which are used by subsequent steps in the
build will allow the codegen script to be more strict and reduce the amount of
configuration necessary.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=143785347