Added `_DEBUG_PRINT_COMPILATION` to offline_compiler.dart.
When set to `true`, outputs the JSON format of the given compile via print()
which is visible in both the transformer and the source_gen compiler frontend.

This is really useful when looking at meaningful diffs of the input.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=147192027