Use top-level `log` instead of `buildStep.logger`
The code that is not importing package:build continues to use a Logger passed
to it. The code that imports package:build uses `log` instead of
`buildStep.logger`. In some places we were passing through the buildStep only
for access to the logger - in those places omit the argument altogether.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=147500615