fix(SourceGen): Re-write NgDepsVisitor to use inputs as source of truth.
Before we checked for reflectables as the indication of whether a
.template.dart file would exist (and we'd need to import and call
initReflector).

After this CL, we use package:build - a combination of both
resolver.isLibrary (to check previously generated files) and buildStep.hasInput
(to check files that will be generated alongside the primary input as part of
this phase).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=148801090