feat(Compiler): Add more initReflector-emitting logic.
This is intentionally not using `code_builder`, since I want to wait for 2.0.0
to land internally and not add migration debt. This won't change significantly
in the future (if anything we will delete code), so as long as it is readable
current structure should be fine.

Partial work towards https://github.com/dart-lang/angular/issues/532.

PiperOrigin-RevId: 164669987