feat(Compiler): Add the beginnings of a GeneratedInjector (emitter+test).
Once hooked up, we will use `@Injector.generate` to create a parallel class as
part of the `.template.dart` generation process. This is an **experimental**
feature.

It currently does not support any of the "hierarchical" flags (i.e. @Self, @SkipSelf), as the intended use case is as part of the bootstrap (root injector) right now. It is relatively easy to support @Optional() in a future commit.

PiperOrigin-RevId: 169469132