Removes unnecessary addition/removal of base href on every navigation
The router implementation checked that each navigation path began with the base
href, then removed it, but every single entrypoint into that code path prefixed
the path it provided with the base href. This change removes both. This is
possibly an artifact of the sharding support that was recently removed.

PiperOrigin-RevId: 178392450