feat(Compiler): Skip <module>.initReflector() by default for `@deferred`.
Before, a `@deferred` component would trigger the AngularDart template
compiler to automatically invoke the component's module's "initReflector()"
method for compatibility with features such as SlowComponentLoader and
ReflectiveInjector.

Going forward for AngularDart v5, we are going to make the defaults suitable
for what is being called "fastBoot", a lighter-weight initialization set and
runtime focused on smaller code-size and faster TTI.

Users may continue to opt-out of "fast_boot" to add the "initReflector()" call.

PiperOrigin-RevId: 179462443