test(Compiler): Add test cases and golden file data for unsupported useValue.
Using @Injector.generate + const Provider(..., useValue: ...) currently is unsupported for almost everything that is not a simple literal value (a string, an int, a bool). The intention here is to support everything already supported by @Component (ViewCompiler), and eventually have the same guarantees for both.

This CL adds data to start iterating on fixes and tracking progress.

PiperOrigin-RevId: 183713604