A number of changes were made here. Let me explain them:
(You can try it here: https://github.com/dart-lang/angular/pull/1084)

* Moved the common application styles into index.html, disabling encapsulation.
  * This loads more styles before the application loads.
  * This allows a sort of SSR-lite, showing the shell of the app before JS loads.
* Prefetches and in-memory-caches the current URL/RPC before the app totally loads.
  * This starts requesting the data from server as soon as `main()` is hit.
  * This avoids any [FOUC][], and makes the application appear to "pop in".
* Replaces pkg/http with plain use of dart:html to drop code-size.

There are probably a bit better ways to do all of this, but it's a start, and was pretty low effort. Other ideas include prefetching the RPC in a 1-line pre-amble JavaScript (even before `main()` is hit), but again, lower impact than this.

[FOUC]: https://en.wikipedia.org/wiki/Flash_of_unstyled_content

Closes #963
Closes #1084

PiperOrigin-RevId: 189658016