Automated g4 rollback of changelist 199385093.
*** Reason for rollback ***

When a `RouterOutlet` is registered for the first time, it navigates to the
current URL with `updateUrl: false`. This is fine if the URL matches a normal
route; however, if that route itself redirects, this change prevented the URL
from updating since the `updateUrl` argument was forwarded.

*** Original change description ***

fix(Router): preserve `NavigationParams` on redirection

Fixes https://github.com/dart-lang/angular/issues/1374.

***

PiperOrigin-RevId: 199494311