refactor(Compiler): emit module URL for style attribution in dev mode
Previously this was emitted in the debug code generation mode, but was emitted
as the empty string in the release code generation mode. With the switch to a
unified code generation mode, we'll now rely on `isDevMode` to tree shake the
module URL away, and pass null instead.

Note this also changed the URL from the template module URL to the component
module URL. While it's orginal purpose was for templates, it's been repurposed
for style attribution, thus the component module URL (which could import
numerous style sheet URLs) is the best fit.

PiperOrigin-RevId: 200603316