Start using IR in offline compiler.
This is just the first checkpoint, to put a stake in the sand, for where we will start the split between the "frontend" and "backend" of the compiler. This is by no means intended to be a final state for any of the parts involved.

RELNOTES=None
(grafted from 4975a9209acd560ded4b24beec21504c653cb253)
PiperOrigin-RevId: 223435324