refactor(Core): remove conditional branch from HTML sanitization
Note the removed comment that suggests `TemplateElement` should be preferred to
`DocumentFragment`. This seems misguided since the implementation of
`TemplateElement` actually relies on a `DocumentFragment` to store its content:
https://api.dartlang.org/stable/2.1.0/dart-html/TemplateElement/setInnerHtml.html

Furthermore, had the condition been met where a `DocumentFragment` was used,
this would have thrown a type error, since the returned inert `Node` was
immediately assigned to a variable of type `Element`. This may have happened to
work in Dart 1, since `DocumentFragment`, despite not implementing `Element`,
happens to implement `innerHtml` and `children`.

PiperOrigin-RevId: 229845502