refactor(*): Remove `AppView.locals` from being set too eagerly.
Now, `locals` is eagerly initialized in the base `AppView()` constructor. As a result, the base constructor is no longer inlined into sub-classes, which looks like a net benefit in any case.

In future optimizations we can remove the field entirely when it is not used (the vast majority of views), but this is quite a bit harder and would not improve code-size (it could even regress code-size a bit in some cases).

PiperOrigin-RevId: 234178281