refactor(Compiler): consistently order tasks for visiting child views
Now when creating a child view, the following tasks are always performed in the
same order:

  * add node to view
  * project node (if necessary)
  * invoke `beforeChildren()`
  * visit all child nodes
  * invoke `afterChildren()`

Note that while this change was supposed to be a no-op, it broke an expectation
in one of our tests, which was very likely testing a bug. The test itself
included a warning explaining the unusual expectation.

Prior to this change, if two nested embedded views each contained the same
directive, and you queried for that directive, the inner directive would be
added first to the query list. This is inconsistent with how queries work in all
other cases, which should keep the directives in the same order as they appear
in the template.

PiperOrigin-RevId: 236390387