refactor(Core): Complete the change from `rootNodes*` to `ViewFragment`.
I needed to add one additional method in order to avoid having to expose the underlying JSArray. I also added a TODO where we could do a bit better, potentially, if we initialized the ViewFragment based on the type of view (better = runtime, not code-size) - but it isn't worth doing immediately.

PiperOrigin-RevId: 240414010