Remove raw types from angular_test.
Congratulations, you are the first code base on which I'm experimenting! (Jenny has already mailed a CL to make DDC's codebase compliant with strict-raw-types.)

This is certainly a lot to review, and by no means do I guarantee it is what you want :P I need to write this all up in a friendly FAQ doc, but here's the big item for now:

Why does `if (v is List)` need to be `if (v is List<dynamic>)`? While it may not read
as well, the idea is that an `is` check is not just a boolean test, it is also _type
promotion_ with meaningful consequences in the type system. Open question? Should
these more-often-than-not be written `if (v is List<Object>)`? Probably not now
because it of perf [1] and it doesn't currently agree with Effective Dart [2]. I
suspect both of these will change in the next few quarters.

[1] https://github.com/dart-lang/sdk/issues/36075
[2] https://www.dartlang.org/guides/language/effective-dart/design#do-annotate-with-object-instead-of-dynamic-to-indicate-any-object-is-allowed

PiperOrigin-RevId: 240834178