Remove unnecessary parentheses.
The UNNECESSARY_PARENTHESIS lint rule will be enforced in google3 soon. This
cleanup prepares this code for that change.

This does not remove _all_ parentheses which are strictly unnecessary; Many
parens which are technically unnecessary are allowed where it is agreed that
they make code more readable in locations of ambiguity. This includes binary expressions (arithmetic, logical, bit, ...), any expression with cascades, expressions with unary ! or -, and more.

Tested:
    TAP for global presubmit queue
    http://test/OCL:240761948:BASE:240750350:1553830978061:ff86a51a
PiperOrigin-RevId: 241011749