refactor(Runtime): rename and document methods on `AppView`
This renames and documents methods on `AppView` associated with inserting into,
moving within, and removing from view containers. The old names were misleading
as they referred to views within view containers as "content children", which is
inconsistent with the use of this term in regards to projected content within
`<ng-content>`.

PiperOrigin-RevId: 242503955