test(Forms): Add an example test of a badly behaving ControlValueAccessor.
**NOTE: This test passes at HEAD due to the bug described.**

Today, NgModel uses ComponentState, which silently does not observe that values "have not changed since last check". This allowed users to implement their own ControlValueAccessor that intercepted "writeValue", and either rejected or transformed the value, writing it back via "registerOnChange".

Ironically, this is what "Expression has changed ..." was supposed to fix, unidirectional synchronous data flow, but it is violated due to a couple of reasons:

1. ComponentState hid the few errors we do throw now (omitted checkBinding).

2. NgModel uses StreamController(sync: true). This was meant to be used for performance only for "bridging APIs" (i.e. change functions that are already asynchronous, like a native `<input>`). In many cases, users do not have asynchronous actions though.

3. ControlValueAccessor does not document or enforce any behavior of its own around proper implementation of writeValue (or whether it is valid to reject or transform on writeValue, for example).

For now, I'd like to check this test in, knowing it will fail if I change NgModel without somehow accounting for this use case. It's possible we just want to go in and update client code.

PiperOrigin-RevId: 242931435