refactor(Runtime): add `RenderView`, and use as type for `AppView.parent`
This interfaces is intended to represent a view that is associated with a
portion of a template and that is responsible for rendering DOM. In the future,
both component and embedded views will implement this interface, while host
views will not.

Unfortunately this requires introducing `unsafeCast(parentView)` in few new
locations to access properties on parent views that aren't general to all view
types.

For example, if a nested embedded view accesses a template local variable from a
parent embedded view, it now must cast its parent view to access `locals`. Note
that we already do this in many places today to access *generated* fields on
parent views, such as references (#) and query dirty flags.

Note this *just worked* before because `AppView` had every field for every view
type. The point of b/129013000 is to specialize the base view type so that this
isn't necessary, but it also makes accessing parent properties more challenging.
The plan is to revisit the necessity of this cast in the future, and see if a
type-based solution could work instead, without affecting code size or
performance: b/130375490.
PiperOrigin-RevId: 243343074