fix(Router): a `popstate` redirect to the active route nows update the location
Navigation requests triggered by `popstate` events that redirect back to the
active route will now correctly update the browser location to match the active
route. Prior to this fix, the browser location would be left in the state
changed by the `popstate` event, since it was assumed that if the target route
was already active, the location would not need to be updated.

PiperOrigin-RevId: 244016986