perf(Runtime): remove parameter of `AppView.initViewRoot()`
This method was always invoked with `AppView.rootEl` as the argument to its sole
parameter. Since this method is on the same class, we can remove the need to
pass the field, and reference it directly within the method.

This change also documents why this method *returns* this field: assigning the
return value of a method you're calling anyways is cheaper than generating
another statement to load a field.

We also reuse the return value of `initViewRoot()` to register event and
`@Output` handlers, instead of relying on the previously cached `rootEl` local
variable.

PiperOrigin-RevId: 244933943