refactor(Test): remove check for condition that is always true
The intent of this check was to ensure the root component under test used
default change detection strategy.

Confusingly, only component views actually encode their change detection
strategy, while embedded and host views are always marked as `CheckAlways` (the
default). Only component view boundaries are used to prune checked portions of
an app during change detection.

Unfortunately this was checking the host view hosting the component under test,
so it will always be `CheckAlways`, regardless of the component.

Correcting this check now introduces failures in client tests, so we'll simply
remove this for now as it's blocking refactors to `AppView`.

PiperOrigin-RevId: 246411817