Implement reading and validation of `@changeDetectionLink` on OnPush components
This annotation will be used to mark an OnPush component as compatible in a
Default app. Such a component will serve as a link between its Default parent,
and a Default descendant that has been loaded imperatively.

When a `@changeDetectionLink` component is visited during change detection,
rather than returning immediately if it's already been checked, it will instead
try to change detect any Default components it has loaded imperatively.

This annotation is not yet publicly visibly, nor are its semantics implemented.

PiperOrigin-RevId: 258265736