refactor(Compiler): remove unnecessary statement visitor
This was primarily used to support event handlers with multiple statements,
which are no longer supported. Now that only a single statement is supported, we
can directly use the expression visitor, and convert the result directly to a
statement.

Note that it's safe to remove the errors thrown for visiting an interpolation or
implicit receiver AST, because the former is enforced when parsing an action by
the expression parser, and the latter isn't possible to reach.

PiperOrigin-RevId: 272899584