Automated g4 rollback of changelist 269459427.
*** Reason for rollback ***

Big uptick in unhandled exception errors since this went in.

The cost of trying to fix this is too high considering we don't know what this was trying to accomplish. We can always try to go down this approach again in the future when we have a design document.

*** Original change description ***

Move ExceptionHandler to angular_compiler.

Also, add handleWarning() to angular_ast's exception handler, so that it can delegate to the new AngularExceptionHandler in angular_compiler.

***

PiperOrigin-RevId: 275127574